feat: create shareable links
