+++ packages/backend/src/graphql/mutations/tiles/index.ts
@@ -1,6 +1,7 @@
 import createRow from './create-row'
 import createRows from './create-rows'
 import createTable from './create-table'
+import createShareableTableLink from './create-table-shareable-link'
 import deleteRows from './delete-rows'
 import deleteTable from './delete-table'
 import updateRow from './update-row'
@@ -14,4 +15,5 @@ export default {
   createRows,
   updateRow,
   deleteRows,
+  createShareableTableLink,
 }
