+++ packages/backend/src/graphql/schema.graphql
@@ -68,6 +68,7 @@ type Mutation {
   ): LoginWithSelectedSgidResult!
   # Tiles
   createTable(input: CreateTableInput!): TableMetadata!
+  createShareableTableLink(tableId: ID!): String!
   updateTable(input: UpdateTableInput!): TableMetadata!
   deleteTable(input: DeleteTableInput!): Boolean!
   # Tiles rows
