+++ packages/frontend/src/graphql/mutations/create-shareable-link.ts
@@ -1,0 +1,7 @@
+import { gql } from '@apollo/client'
+
+export const CREATE_SHAREABLE_TABLE_LINK = gql`
+  mutation CreateShareableTableLink($tableId: ID!) {
+    createShareableTableLink(tableId: $tableId)
+  }
+`
